Extract default datatypes from compilers
